% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_p_curves}
\alias{compute_p_curves}
\title{Iterative p-values
Computes iterative p-values}
\usage{
compute_p_curves(df, n_iter, fct_levels)
}
\arguments{
\item{df}{input dataframe from prepare data function}

\item{n_iter}{number of iterations}

\item{fct_levels}{optional}
}
\value{
dataframe "all_iterations"
}
\description{
Iterative p-values
Computes iterative p-values
}
