% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{p_plot}
\alias{p_plot}
\title{Parameters and Plot
Gives iterative-p-value plot and metrics}
\usage{
p_plot(n_iter)
}
\arguments{
\item{n_iter}{number of iterations}
}
\value{
parameters and plot
}
\description{
Parameters and Plot
Gives iterative-p-value plot and metrics
}
