% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_data_against}
\alias{prepare_data_against}
\title{Prepare data (against each other)
#' Prepares dataframes for testing against each other}
\usage{
prepare_data_against(df_input, factor, var, id)
}
\arguments{
\item{df_input}{input dataframe}

\item{factor}{two-staged factor for testing against each other}

\item{var}{dependent variable}

\item{id}{unique identifier}
}
\value{
prepared dataframe
}
\description{
Prepare data (against each other)
#' Prepares dataframes for testing against each other
}
