% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_data_zero}
\alias{prepare_data_zero}
\title{Prepare data (against zero)
Prepares dataframes for testing against zero}
\usage{
prepare_data_zero(df_input, var, id)
}
\arguments{
\item{df_input}{input dataframe}

\item{var}{dependent variable}

\item{id}{unique identifier}
}
\value{
Prepared dataframe
}
\description{
Prepare data (against zero)
Prepares dataframes for testing against zero
}
