% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_p_values}
\alias{summarize_p_values}
\title{Summary
Summarizes information about iterative p-value-function}
\usage{
summarize_p_values(df_iterations)
}
\arguments{
\item{df_iterations}{input dataframe from compute_p_curves function}
}
\value{
dataframe "summary_stats"
}
\description{
Summary
Summarizes information about iterative p-value-function
}
